# api/docs/swagger.yaml
definitions:
  models.Branch:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Bucket:
    properties:
      branch_id:
        type: string
      client_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      updated_at:
        type: string
    type: object
  models.Category:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      photo:
        type: string
      type:
        items:
          $ref: '#/definitions/models.CategoryPrduct'
        type: array
      updated_at:
        type: string
    type: object
  models.CategoryPrduct:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      photo:
        type: string
      type_id:
        type: string
      updated_at:
        type: string
    type: object
  models.Client:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.CreateBranch:
    properties:
      name:
        type: string
    type: object
  models.CreateBucket:
    properties:
      branch_id:
        type: string
      client_id:
        type: string
      product_id:
        type: string
    type: object
  models.CreateCategory:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      photo:
        type: string
      type_id:
        type: string
      updated_at:
        type: string
    type: object
  models.CreateClient:
    properties:
      name:
        type: string
    type: object
...
  models.Empty:
    type: object
  models.GetBucketByClientID:
    properties:
      branch_id:
        type: string
      client_id:
        type: string
    type: object
  models.GetBucketByClientResponse:
    properties:
      total_price:
        type: integer
    type: object
  models.GetListCategoryResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.CreateCategory'
        type: array
      count:
        type: integer
    type: object
...
  models.Product:
    properties:
      branch_id:
        type: string
      category_id:
        $ref: '#/definitions/models.CategoryPrduct'
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      photo:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  models.UpdateCategoryPut:
    properties:
      branch_id:
        type: string
      category_id:
        type: string
      name:
        type: string
      photo:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /branch:
    post:
      consumes:
      - application/json
      description: CreateBranch
      operationId: CreateBranch
      parameters:
      - description: CreateBranchRequestBody
        in: body
        name: Branch
        required: true
        schema:
          $ref: '#/definitions/models.CreateBranch'
      produces:
      - application/json
      responses:
        "201":
          description: GetBranchtBody
          schema:
            $ref: '#/definitions/models.Branch'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: CreateBranch
      tags:
      - Branch
  /bucket:
    post:
      consumes:
      - application/json
      description: CreateBucket
      operationId: CreateBucket
      parameters:
      - description: CreateBucketRequestBody
        in: body
        name: Bucket
        required: true
        schema:
          $ref: '#/definitions/models.CreateBucket'
      produces:
      - application/json
      responses:
        "201":
          description: GetBuckettBody
          schema:
            $ref: '#/definitions/models.Bucket'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: CreateBucket
      tags:
      - Bucket
  /casier:
    post:
      consumes:
      - application/json
      description: CreateCasier
      operationId: CreateCasier
      parameters:
      - description: GetBucketByClientIDBody
        in: body
        name: Bucket
        required: true
        schema:
          $ref: '#/definitions/models.GetBucketByClientID'
      produces:
      - application/json
      responses:
        "201":
          description: GetBucketByClientResponseBody
          schema:
            $ref: '#/definitions/models.GetBucketByClientResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: CreateCasier
      tags:
      - Bucket
  /category:
    get:
      consumes:
      - application/json
      description: Get List Category
      operationId: get_list_category
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCategoryListBody
          schema:
            $ref: '#/definitions/models.GetListCategoryResponse'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get List Category
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create Category
      operationId: create_category
      parameters:
      - description: CreateCategoryRequestBody
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryPut'
      produces:
      - application/json
      responses:
        "201":
          description: GetCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create Category
      tags:
      - Category
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Category
      operationId: delete_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: DeleteCategoryBody
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete Category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get By ID Category
      operationId: get_by_id_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get By ID Category
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update Category
      operationId: update_category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateCategoryRequestBody
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryPut'
      produces:
      - application/json
      responses:
        "202":
          description: UpdateCategoryBody
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update Category
      tags:
      - Category
  /client:
    post:
      consumes:
      - application/json
      description: CreateClient
      operationId: CreateClient
      parameters:
      - description: CreateClientRequestBody
        in: body
        name: Client
        required: true
        schema:
          $ref: '#/definitions/models.CreateClient'
      produces:
      - application/json
      responses:
        "201":
          description: GetClienttBody
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Invalid Argumant
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: CreateClient
      tags:
      - Client
...
swagger: "2.0"